LayoutTests/platform/mac-catalina/fast/text/international/system-language/navigator-language/navigator-language-pt-BR-expected.txt

FAIL navigator.language should be pt-BR. Was pt.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
